CMS-13590 [Backoprt 14.3] Support XPageLayout via HST and JcrTemplateNode in HST Model
Originally the plan was to get the available XPageLayout instances for
an Experience Page folder via getting hold of the right Channel and on
the Channel object we would get #getXpageLayouts.

However, this turned out to be problematic: Namely, all Channel objects are
loading actively when loading the HST Model, BUT the Hst Configurations
(HstSite) PER Channel are loaded lazily and on request. However, the
XPageLayout objects would need the 'hst:xpages' section from the hst
channel configuration to be loaded, which is however only loaded as part of
the HstSite. Storing the XPageLayout objects in a Channel object could thus
only be done by killing the lazy loading of HstSite objects, which is not
an option.

A second problem was that typically the cms user adding a new XPage document
is not allowed to read the hst XPage layout configuration nodes at all. That
would then have to be done by some other jcr session.

Also, during a single request, all objects resulting from hst configuration
should have the lifecycle of the HST Model itself

For the above reasons, it was clear that the way to go forward was:

1) Not exposing XPageLayout objects any more via Channel since simply not
possible
2) During Hst Model HstComponentConfigurationService instance loading,
directly for XPage layout pages load the JcrTemplateNode for such an
XPage HstComponentConfigurationService instance and make sure this
JcrTemplateNode is sealed (aka immutable)
3) To the ChannelService, add the method

Map<String, XPageLayout> getXPageLayouts(String channelId);

which gives the CMS the possibility to invoke the following code:

final ChannelService channelService = HippoServiceRegistry
                   .getService(PlatformServices.class).getChannelService();
final Map<String, XPageLayout> xPageLayouts = channelService.getXPageLayouts(channelId);

Only TODO is to support mixins in JcrTemplateNodeConverter which requires
mixin support on HstNode

(cherry picked from commit 5ce99006fbbb2be4cde0a2367e0f114a89e65f17)