CMS-12058 Expose field values of compare document
Added a new method #getFieldCompareValue that is similar to
#getFieldValue apart from the fact that it reads the field values from
the compare model.

Doing a lookup of the compare model (the complete document) is a bit
tricky as FieldPlugins normally only require access to their own compare
value, which is handled by the editing framework. To support this, the
ComparePlugin has a new public method which return an IModel<Node> that
contains the full document that we want to compare against.