TRIVIAL [Backport 14.3] invalidate http session: jcr session must be logged out
In case an authenticatedCmsUser is send with the request, the
CmsSecurityValve creates a jcr session which gets tied to the http
session. If the session does not get invalidated, the session is kept
in memory until the http session gets GC-ed (because test has passed)
and then, a stack trace will be logged during tests that a jcr session
was not logged out. Hence invalidate the http session at end of the
request during testing

(cherry picked from commit c88efdb15812c76a6b11fdcd3f849f0aa7309d5b)