CMS-13647 Enable more wf actions for transferable
Enable copy, move and delete in the case there only is a transferable
draft variant.

Some test cases become obsolete, because a document can only be
transferable until the first save action, so all cases for transferable
documents with a published or unpublished variant are obsolete.

Make the CopyDocumentTask copy documents with only a draft variant as
well. In that case the draft variant is copied.

(cherry picked from commit 96700c3abf21637155396d5084261de3db783880)