CMS-13693 Associate revision with handle
Associate revision with handle as well in the HandleIdentifierStrategy,
because revisions should also be handled by the
HandleIdentifierStrategy.
As the DefaultCmsEditor and the HippoStdPublishableEditor both generate
usage events on render, the DocumentUsageEvent should get the handle
identifier from the HandleIdentifierStrategy not only for documents and
handles, but also for revision nodes, because these two editors also
are viewers for revisions.
CMS-12441 introduced this bug by only dealing with documents and
handles. Now that this is resolved the warning that was logged in the
console not does not appear anymore and the usage events for versioned
nodes are associated with their associated handle.

(cherry picked from commit b59981b0a95c5d972460a76ab295de4d6d245f32)