CMS-13647 Skip archive if no unpublished variant
Skip archive if there is no unpublished variant.

Delete is now enabled for a transferable draft (A draft that is saved as
draft).
When a document is deleted it's archived (See delete transition in scxml).
The ArchiveDocumentTask archives the unpublished variant.
In the case that there is no unpublished variant a NullPointer is
thrown.
So a NullPointer is throws when deleting a transferable document.

When the archivation is skipped if there is no unpublished variant a
NullPointerException will not be thrown.

(cherry picked from commit 8b386b849d54fc5bb2fb35bb8ddd30a9c9a5a009)