CMS-13700 Use BranchHandle for info message
Replace determining draft changes, unpublished changes and live/offline
by using the BranchHandle and the draft variant instead of the hints  so
authors will see the correct state of the document.

Authors saw the document was unpublished when the state of the document
was published in the info (yellow ribbon).
To determine if a document is published or unpublished the "publish" hint was used.
Authors are not allowed to publish a document, so the info said the
document was offline while it was live.

(cherry picked from commit 68ba1050403f16d1af1f3ade17a04fb07d67dee4)