CMS-13701 Add canKeepDraft to document info
Add canKeepDraft to document info, because the frontend only has to
render the save draft button in the second step of the document creation
process in the visual editor if the document workflow contains a hint
"saveDraft" that is Boolean.TRUE.

The frontend should only show the button in that case, because if the
project contains custom scxml, the save draft feature should not be
enabled, unless the scxml is updated with the chances necessary for save
draft. See
https://issues.onehippo.com/browse/CMS-13647?focusedCommentId=579605&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-579605.

So add the canKeepDraft parameter to the document info, so the visual
editor can determine if the save draft button should be shown.

For verification, observe the returned value of the rest call, change
the scxml so that the saveDraft hint is always false to see that the
button won't be added in case the scxml changes were not incorporated.

(cherry picked from commit 4aec65ce87bef6f5ce93fb745332fd43d9350b9c)