CMS-13701 Add the save draft button
Add the save draft button if the document.info.canKeepDraft is present
and true.
The logic of the save draft button is exactly the the as in
editContentMain and also uses the ContentEditor to handle the saving.

(cherry picked from commit 0e911bfd09b3181f613cbe5cda1009d47c394b64)