CMS-13515 Hide layout selector for new folders
When a user creates a new folder that is a child of and xpage folder the
new folder inherits the XPage layout of the parent. ( specs )
When a user creates a new folder that is a child of a (non) xpage folder
the new folder is also a (non) xpage folder.
So when a user creates a new folder the XPage layout selector should
never be visible.

The prototypesAndNames parameter of
org.hippoecm.frontend.plugins.standardworkflow.AddDocumentDialog#hasFolderPrototype
has been filled in
org.hippoecm.repository.standardworkflow.FolderWorkflowImpl#prototypes.
The prototypes are added to the hints.
The documentType of a new folder or directory is hippostd:folder or
hippostd:directory.
For those document types the type is added in
FolderWorkflowImpl#prototypes.
So if the added node is a folder or directory the type is an element of
the prototypesAndNames set.

Determine the visibility of the XPageLayoutContainer is one place
instead of two places to improve readability.

(cherry picked from commit 82deac87d73cfb450efaf108d16d071b446233cc)