CMS-13648 Prevent pink boxes before message
org/hippoecm/addon/workflow/ConfirmDialog.html defines a
wicket:enclosure for the text field. An enclosure adds a transparent
container that's not part of the markup ( see
https://ci.apache.org/projects/wicket/apidocs/6.x/org/apache/wicket/markup/html/internal/Enclosure.html
). The "notification" css class matches the color of the boxes.
Using the intro instead of the text resolves the issue. I've created
RD-5615 to investigate and solve the issue as it's out of scope for this
issue.

(cherry picked from commit 49a7d7026a66d2db51d82e2807f2660783d93c22)