CMS-13666 Add DocumentWorkflowInvokerPlugin
The DocumentWorkflowInvokerPlugin is used to trigger document workflow
actions from javascript. This means it should work without a document
being selected in the the content perspective.

It uses the MenuHierarchy to locate the desired submenu or category and
in the submenu tries to find the required action.

The plugin will register three new JS functions in global scope:
 - Hippo.Workflow.invoke(documentId, category, action): Promise
 - Hippo.Workflow.resolve(payload)
 - Hippo.Workflow.reject(payload)

These are used to #invoke a new workflow action and subsequently either
#resolve or #reject that action.

(cherry picked from commit b541efb136581b87bfb3e7f305136e66716b539b)