CMS-13666 Expose XPage actions
- publish
- schedule publication
- request publication
- request schedule publication

- unpublish
- schedule unpublication
- request unpublication
- request schedule unpublication

- copy
- move
- delete

It is assumed that if a user is allowed to publish/unpublish, the user
is also allowed to schedule a (un)publication. Likewise, if a user is
allowed to request a (un)publication, the user is also allowed to
schedule a request for (un)publication.

(cherry picked from commit c2e4c9ea9acb9b879b5ee6589b8aa23d8cce24ae)