CMS-13726 Prevent warning in logs
Prevent warning in logs during execution of tests. The warning is logged
in case a document type cannot be found. The log level in that case
should at least be warn or error. A warning is logged before the
exception, because the exception mechanism used only allows a strict set
of messages, because the exceptions have to be translatable to be used
in the frontend.
The test should trigger this specific piece of code, so the only way to
prevent the warning is to increase the log level for this specific
class.

(cherry picked from commit 77cf348fa59b003a33801be208e95338abe47ec9)