CMS-13727 Prevent costly session refresh
CMS-13358 introduces a session refresh when switching from visual to
content editor and vice versa.

The reason to refresh the session were warnings that the transferable
property of the draft document was out of sync. That was caused by
faulty use of scxml ( See CMS-13409 ).
CMS-13409 removes on-entry modification of properties of the draft
variant.
So the "transferable" property is not out-of-sync anymore.
So the session refresh can be removed.

The removal of the session refresh did not cause the warning again.

(cherry picked from commit 30bc37fbf394cb9486e58468b808d2baa384c0bf)