CMS-13237 Support creating manually a new version for a (branch) document
Via a POST REST call to

/cms/ws/content/workflows/documents/{handleId}/version

you can now trigger a new version for the *currently* selected branchId
on the cms session context (aka the branch the cms user is looking at
in the CM).

The implementation is not ideal since this is meant for the 14.3 and
means we are not allowed to modify scxml, meaning it is too much of a
hassle to introduce the workflow method

documentWorkflow.version(branchId)

which should do the equivalent of

if (!documentWorkflow.listBranches().contains(branchId)) {
    throw new ForbiddenException(new ErrorInfo(Reason.BRANCH_DOES_NOT_EXIST));
}
if (Boolean.TRUE.equals(documentWorkflow.hints(branchId).get("checkoutBranch"))) {
  documentWorkflow.checkoutBranch(branchId);
}
documentWorkflow.version();

If we add it now to the document workflow, we need to implement it
differently in 14.3 without scxml changes. This is too much
hassle for now, to be picked up later

(cherry picked from commit 594d61af3d01108e63ed945b39180d913ca803d7)