CMS-13702 Combine states and actions
Computation of state usually requires similar data as
computation of the action flags. So to reduce code duplication
it is best to implement state and action computation in one
service. So instead of having an ActionProvider and a StateProvider
we now have a ComponentInfoProvider that returns an (optional)
ComponentInfo.

In the wpm module the WpmComponentInfoProvider implements the
interface. It provides the project related info, like review
state of channel and xPage, and the project name. Note that
the WpmComponentInfoProvider returns an empty option for the
master branch. Therefore, we do NOT have to localize the project
name. That would have been quite a hassle because the
ProjectLocalizationService is not readily available from the
wpm/hst/platform module.

I decided to name it ComponentInfo because it is used in the ComponentResource,
which deals with hst:component uuids. I think a better name would be
ActionStateInfo so I'm going to do a rename in a next commit.

(cherry picked from commit f7e1d3a3e956ab9796b7b4a9c6062b8052458e77)