TRIVIAL Fix template interpolation issues
Considering the documentation of angular.js for ng-value it requires the value to be a string. My guess is that with angular, where [ngValue] can be an object, this API 'breaks' between angular.js and angular. The ngValue will now be parsed handled by Angular instead of angular.js as its loaded later.

(cherry picked from commit 971599400380509e2562c11b274c503373afff7b)