TRIVIAL Fix template interpolation issues
Considering the documentation of angular.js for ng-value it requires the value to be a string. My guess is that with angular, where [ngValue] can be an object, this API 'breaks' between angular.js and angular. The ngValue will now be parsed handled by Angular instead of angular.js as its loaded later.

(cherry picked from commit 4e91bc220ceb7284dc8ff72128e5170458627833)