CMS-13640 [Backport 14.3] Support XPage documents in SiteMap in Channel Mgr
By default from now on, XPage documents are shown in the CM SiteMap.
Important detail is that a 'delegated jcr session' is used which is a
combination of the access of the preview hst session and the cms user
session : even an XPage which the cms user does not have read access on
will be visible in CM SiteMap if the preview hst user can read the XPage
document. At the end of the request, the Cleanup valve will logout the
delegated session

Implementation details:

- In case there is an explicit sitemap item config for an XPage Document,
the SiteMap Item config is used and the XPage document is not added
(since already represented)
- The XPages are appended to the List of the Sitemap Items, so added at
the end
- The XPages are just like SiteMapItem Pages sorted on the 'pathInfo'
- The XPage documents will have the same name as the items in the Console
(translated name)
- For customers who do not want the XPage in the sitemap (for example
because they have thousands of them, destroying performance and it is
not manageable, I have added an hst-config property to be able to switch
off XPage Documents in the sitemap: this can be done in an end project
by setting :

channelmanager.sitemap.hide.xpages = true

(cherry picked from commit a92fdf5e6428f8c28fb78c79987a03fc029f59d3)