CMS-13666 Add XPage actions
- publish
- schedule publication
- request publication
- request schedule publication

- unpublish
- schedule unpublication
- request unpublication
- request schedule unpublication

- copy
- move
- delete

It is assumed that if a user is allowed to publish/unpublish, the user
is also allowed to schedule a (un)publication. Likewise, if a user is
allowed to request a (un)publication, the user is also allowed to
schedule a request for (un)publication.

(cherry picked from commit 4603ef9c1570c78f0a2a72d567b97d9b7fac2f20)