CMS-13666 Add XPage actions
- publish
- schedule publication
- request publication
- request schedule publication

- unpublish
- schedule unpublication
- request unpublication
- request schedule unpublication

- copy
- move
- delete

It is assumed that if a user is allowed to publish/unpublish, the user
is also allowed to schedule a (un)publication. Likewise, if a user is
allowed to request a (un)publication, the user is also allowed to
schedule a request for (un)publication.

(cherry picked from commit 84d5833bc6f3e72d1064c3c67f5e466cc03b557f)