CMS-13701 Do not discard changes on keep draft
Do not call ContentEditor.discardChanges, so the error "Document already deleted" is not thrown
so the discard method is not called,
so the document is not deleted,
so the error is not thrown.
In the second step of document creation there is only a draft variant
If there is only a draft variant, discard deletes the document
On KeepDraft the dicard ContentEditor#discardChanges was called
So the document was deleted
so an error was thrown.

(cherry picked from commit 6eb2d254bc57fe51836bcef25f87ba9454e75e18)