CMS-13701 Enable and show the "Save draft" button
Show the save button when the draft is retainable
so a "retainable" document can be saved without a change.

In the steps after content creation show the save draft button
when the draft is retainable, because save draft is only allowed for
new documents ( only a draft variant exists ).

(cherry picked from commit 27943f6bbfcaa62927538adbac8228e1add61485)