TRIVIAL Avoid Spring initialization errors
Since the 'components/tests' module pulls in the 'client' module,
and the client module has a Spring 'module' extension, that Spring
module gets loaded as well during the 'components/tests' tests.
The client Spring module has a reference to
ref="jcrObservationEventListenerItems"
but the 'jcrObservationEventListenerItems' is part of the
SpringComponentManager-observation.xml which is the only spring xml
which is not included during tests since it leads to all kind of
corner cases. In AbstractTestConfigurations.xml it can be seen that
SpringComponentManager-observation.xml is missing.
Hence to AbstractTestConfigurations.xml, we can add an empty
jcrObservationEventListenerItems list to avoid the spring wiring
errors being logged

(cherry picked from commit f39c657b8e92b745bce46b770cd0ecb56fd02059)