CMS-13673 Allow authors to commit editable instance
If a document has auditTrace enabled then a version must be saved in version
history. This can easily be implemented in scxml, but that is not possible in
a minor release. Unfortunately DocumentWorkflowAction.version()) requires at
least editor privileges, so for authors we can't trigger this action inside
of commitEditableInstance. Way around this is to instantiate and execute a
VersionVariantTask.