CMS-13819 Hide restore buttons by inspecting hints
According to SCXML authors are never allowed to restore a version.
Therefore the hints "restoreVersion" and "restoreVersionToBranch"
are always false for authors. However, the UI does not inspect the
hints of a document to decide if the "Restore" button must be visible
or not.

Note: I added the field documentHintsGetter (a BiFunction) to the
DocumentVersionServiceImpl to avoid static method mocking nightmares.

(cherry picked from commit 8ab858d93cfb91c258ebf387c13a3406a9bd7db0)