CMS-13818 Switch perspective after ve->ce switch
The navapp calls beforeNavigation on all connected app before switching
to a new perspective.
The cms is on of the apps.
The cms call beforeNavigation on all its perspectives.
The channel manager perspective is on of those perspectives.
So the navapp call beforeNavigation when the user want to switch to another app.
If one of the returned promises does not resolve, the navapp does not switch
perspective.

The beforeNavigation call closes the channel manager.
When the channel manager close the canUiExit callback is called.
In the case of a retainable document that method calls saveDraft.
The document has already been closed, so document is undefined.
The keepDraft method sets the retainable property of the info property
of the document.
If a property is accessed on an undefined object javascript throws an
error.
So the beforeNavigation call is cancelled so the perspective does not switch.

The keepDraft call is unnessary (introduced as part of CMS-13462, which
is obsolete, see issue description). As is the additional call in the
OpenDocumentListener.

Let the uiCanExit in case of a retainable document return a resolved
promise and close the right side panel so the beforeNavigation call
resolves so the perspective is switched.

Also check if the document is defined before accessing the info property
and otherwise skip the backend call.

The uiCanExit method should return a promise as pointed out by Arthur
Bogaart.

(cherry picked from commit 59eaa9b53c6cc30c02c4062b2ed8309b1c57b46a)