CMS-13655 Show the correct state icon and info
Show the correct state icon and info message in the content editor after
switching from channel manager to content editor.

At the switch from the channel manager to the content editor the
document is discarded after saving by the channel manager.  See
community/channel-manager/frontend-ng/src/app/channel/sidePanels/
rightSidePanel/editContent/editContentMain/
editContentMain.controller.js:191
In the scxml the retainable property was set to true for the
disposeEditableInstance state.
When a document is retainable the icon and info messages associated with
a saved draft is displayed.
So when a document is discarded the icon and info displayed the wrong
state.

Set the retainable property to false on disposeEditable instance,
because in the latest specs save draft should only be available for new
documents. DisposeEditable instance requires an unpublished variant, so
when disposing there will never be a retainable document.

(cherry picked from commit 43ddb5a64581f1f6169a854befe4e783d228720a)