CMS-13480 [forward port 14.4] Support switch template for XPage documents as well
First of all, there were two issues with the PropertyRepresentationFactory :

1) It relied on final ContainerItemHelper containerItemHelper while for XPages
the containerItemHelper does not add anything since it is really there for
supporting locking of containers which is not desired for XPage documents
2) It relied on HstComponentParameters while XPage documents have
XPageComponentParameters. Better to use the common superclass
AbstractHstComponentParameters

Therefore, modify the interface to have args

  final HstComponentConfiguration componentConfiguration,
  final AbstractHstComponentParameters componentParameters,

instead of

  final ContainerItemHelper containerItemHelper,
  final HstComponentParameters componentParameters,

Since the PropertyRepresentationFactory is part of page-composer and no
api, this change is allowed

After making this change, it is clear that the
XPageContainerItemComponentServiceImpl
should gets its dependency on the containerItemHelper dropped:
it shouldn't be used anyway for XPage documents as mentioned above.

Finally, as a result, it follows that
ParametersInfoProcessor#getPopulatedProperties
required the HstComponentConfiguration as argument, since the
HstComponentConfiguration can either be backed by hst configuration
component item or by an XPage document for the request created
HstComponentConfiguration instance. Also it turned out that the
complex and brittle logic in ContainerItemComponentRepresentation:

// TODO Below is EXTREMELY tricky and actuall brittle: from a componentItemNode you do not have enough information
// TODO to backtrack the HstComponentConfiguration instance that was clicked on in the Channel Mgr: Namely, a
// TODO component item can of course be inherited by a news page as well as an agenda page, for example if the
// TODO component item is in the inherited 'header' container in abstractpages. So, below you just happen to get
// TODO a component reference instance, not perse the one clicked. Then again, if you do not call #getParent
// TODO in general it most likely will go fine
final HstComponentConfiguration componentReference = componentConfigurations.values().stream() //TODO SS: Optimize loop
     .filter(x -> x.getCanonicalStoredLocation().equals(componentCanonicalPath)).findFirst()
     .orElseThrow(() -> new RuntimeException(String.format("Component representation is not found: '%s'",
     componentCanonicalPath)));

could be simply replaced with

final HstComponentConfiguration configObject =
            containerItemHelper.getConfigObject(componentItemNode.getIdentifier());

The rest of the changes are to fix mocks in tests

(cherry picked from commit d8e8286f4aae3e18dd0d45899f72f379abb16376)
(cherry picked from commit bfa5b3352942dd853f44fa9c0c5aa4e7b57258d2)