ENT-1989 [Backport 14.3] Lock XPage document if not branched and viewing branch in CM
When in the CM a user looks at branch 'foo', and navigates to an XPage,
then when that XPage is not branched for 'foo', it should not be
modifiable in the CM : only after it being part of project 'foo' it
is allowed to modify the XPage as part of branch 'foo'. This is fully
similar to content editing of documents which are not part of the branch
being looked at.

For the implementation, this means that we need to check the hints wrt
the selected branch on the CmsSessionContext. To avoid code duplication
the logic to find the current selected branchId in the XPageUtils has
moved to HstRequestUtils

(cherry picked from commit 378a0c004fe76246ea1ec23c9bd6ffb7ca235b81)