ENT-1945 [Backport 14.3] Support correct feedback message for locked component item
Note: translations of the feedback message has not been taken care of

A component item from an XPage document can be locked for a plurality of
reasons, for example because

1. someone else is editing the document
2. there is a scheduled (de)publication
3. the document is not part of the branch that is viewed in the
channel manager
4. the document is being part of a branch that is under review

Specifically, case 4, the reason can only be known by a downstream
project. Hence introduced the concept of

cmsComponentAttributeContributors

in CmsComponentWindowResponseAppender which can try to find a reason
for why the document is being locked. The first
ComponentLockedReasonContributor that returns a reason is the reason
that is being set for the lock icon in the Channel Manager

(cherry picked from commit 6af717562e1420ee175473284c7b1aa4fc87a77c)