CMS-13367 [Backport 14.3] Mark container(item) HST-XPage-Editabl = true in case editable
Even if the channel is under reviewm, a rejected XPage should become
editable again. The CM frontend code should instead of globally lock
the CM overlay in case the channel is under review, check container (item)
html comments to inspect whether the overlay should be re-enabled because
the XPage should for example be editable again after it has been rejected
in in a review process for example

Since the overlay can be enabled again for the XPage containers, we have
to take care that HST configuration elements do not become editable
again: container (items) from hst configuration already have the
locked by system marker in case of a review for the channel. Now also
made sure that the 'edit menu' button is not shown when the overlay is
enabled again for an XPage but the Channel is still under review (and
not rejected)

(cherry picked from commit 096d0dd33d258860a2fb043d9745a47bee5c3f48)