CMS-13519 [Backport 14.3] Optimistic locking for XPage Documents in Channel Manager
the feedback for the current solution will be the same as the
already modified' for normal hst configuration changes: Unfortunately,
there are two minor things with that

1) The message only says the action failed, not 'Another user has made
concurrent changes. Reload the page'.
2) The page is not automatically reloaded in case of a failure response

For now, it works the same as it was for normal hst config changes. Also
note that in general, the CM should reload the page on an error response
from the backend, so number 2 should be picked up more generically. For
that there is CMS-13844

For now, thus the same scenarios are tested as for normal hst
configuration.

Unfortunately, adding a component item to a container does not send the
timestamp, so for adding a container item, we do not check whether
someone else already modified the container (after the page was loaded
in the CM). This is acceptable though, since adding a container item can
just be done successfully even if another user has modified the
container concurrrently.

Actions which are checked with optimistic locking are

1) modifying a container item
2) drag dropping a container item within a container
3) drag dropping an item between two containers

Creating experiments is also validated to work correctly with the
optimistic locking : it behaves a bit different wrt locking when
starting an experiment (via publish document) but it behaves just fine
and without locking issues

Implementation details:

- In JcrUtils, made the isAncestor method public since handy for many
other classes as well.
- The 'XPageUtils#validateTimestamp' for XPages works a bit different
than for hst config pages: Since hst config pages use locking info, but
since XPages does not have this, we use 'lastModifiedBy' for XPage
documents: Ideally this would not be needed but the Frontend does
sometimes do multiple consecutive requests without updating the
timestamp : on the backend validating the timestamp only would fail
always for the second request
- in XPageUtils#checkoutCorrectBranch when checking out a branch, the
branch should only be checked out *if* and only *if* it is the latest
version of the branch: otherwise someone else already did modify the
XPage document for that branch. If we blindly restore the versioned
xpage doc branch, then the 'validateTimestamp' would always pass since
also the timestamp to validate against would be restored. Hence, only
checking out the latest version of a branch is allowed.
- For XPage document request based HstComponentConfiguration instances,
the containers are partly cloned from the XPage Layout and partly from
the XPage document : however, the timestamp was not coming from XPage
Document but from the layout, which was of course incorrect, therefor
adjusted this in HstComponentConfigurationService

(cherry picked from commit c56911491637397539eec64062f55c363e1ea7aa)