CMS-13367 [forward port 14.4] Enable overlay for rejected XPage content
The overlay toggle buttons should be evaluated on every page change.

If we are on a project branch and we can not edit the project, we do an additional
check to see if there is a container on the page marked as HST-XPage-Editable=true.

If so, we enable the toggle buttons.

(cherry picked from commit 3ea18de9a92c164da59a54fc7c913a80371af2e5)
(cherry picked from commit 09dc1f30f3e0e7ffa277aba62c08e37b90fb61e2)