CMS-13759 Preserve breadcrumb on open document
Update the navLocation when the editor has succesfully been closed, so
the breadcrumb will show the current channel and the channel manager
menu item in the navapp is selected.

The breadcrumb is first updated when the content editor is closed and
then immediately updated again when the channel manager frontend-ng code
is notified. The change is not noticeable.

[frontend-ng]
When the channel manager opens a document, it will publish a
"close-content" (ContentEditorService.open).
[frontend]
The Ext-js part subscribes to that event and fires a "close-document" event.
The associated java class: ChannelEditor.java creates a new
CloseDocumentEditorEventListener that closes the editor if exists and is
valid.(org/onehippo/cms7/channelmanager/channeleditor/CloseDocumentEditorEventListener.java:79).
[cms]
If the editor closes it nullifies the backing node of the documentModel.
The BrowserPerspective listens to changes in de documentModel and
updates calls navapp.updateNavLocation when the model changes.
When this editor is succesfully closed, a channelEditor.closeDocumentResult is called
by the CloseDocumentEventListener.
[frontend]
The channel.editor publishes a "close-content-result" event.
[frontend-ng]
The cms.service subscribes to that event and the promise will resolve once that event is received.

The updateNavLocation call upon model disposal is needed to remove the
last part of the breadcrumb when the document tab is closed in the
content editor.

(cherry picked from commit 5f1125f53b3df3528e79eab7639a29d11b167d12)