CMS-13822 Pure refactoring
The xPageMenu service only swithes to the new 'edit-page-content' state.
It does not take into account to switch to another project when the
document is loaded again.
It should. To be able to this refactor the function to take an optional
state argument.

(cherry picked from commit 67fff2d3b6108e8bf8b7bc1aef41740034bd1ae4)