CMS-13822 Close editor on project switch
Add the new states to the states for which "editing" is true:

The xpage feature introduced two new states 'edit-page.content' and
'edit-page.versions'.
Before switching project a check is made if the editor is open. That
check checked for the state 'edit-content'.
On before switch the editor is closed.

Enter the state to add a document to the project:
The xpage feature immediately switched to the new states, without
checking if there was a branched version of the document.
That caused the document to be opened for the wrong branch.

Showing the versions follows the same logic.

When a user in the channel manager tries to edit an xpage document that
is not yet associated with the same project as currently selected, then
the visual editor goes to a state where the user is asked to add the
document to the project.
After adding the document document is shown in layout without versions.

Distinguishing between xpages and normal documents is too complex for
the scope of this issue. See
https://issues.onehippo.com/browse/CMS-13879.

(cherry picked from commit d082eae9a54347e2eadc411686627bb870655f9b)