CMS-13854 Only use unpublished xpage branch if equal to selected one in XM
If the branchId of the unpublished variant is different from the selected one
in XM and we would use that one to compute the x-page state and actions then
the user would see state and actions of a non-selected branch, which is very
confusing. So we should fall back to master branch if unpublished and selected
branch are different.

The x-page rendered in the XM is not always equal to the workspace unpublished
variant (the one under the handle). This is e.g. the case if an x-page is part
of a project while somebody is editing the core variant. So we should always
use the UUID stored in the hst context as attribute
REQUEST_EXPERIENCE_PAGE_UNPUBLISHED_UUID_VARIANT_ATRRIBUTE

(cherry picked from commit d95b3e2ec7becfd56e60fa0ac74a754517eed76c)