CMS-13889 [Backport 14.3] Serialized JcrPath based params as-if they are HippoBean's
Instead of serializing something like

"meta" : {
        "paramsInfo" : {
          "path" : "common/contact"
        },
        "params" : { }
}

we want

"meta" : {
        "paramsInfo" : {
          "path" : "/page/u30092f4e2ef74c7286a58ce895908937"
        },
        "params" : { }
      },
...
...
...
"u30092f4e2ef74c7286a58ce895908937" : {
      "type" : "document",
      "links" : {
        "site" : {
          "href" : "/about-us",
          "type" : "internal"
        }
      },
     ....

To be able to achieve this, use a custom
PageModelApiParameterValueConvertor which does specific handling for
parameters backed by JcrPath : for such parameters, an attempt is made
to find the referenced document, and if found include that in the
serialized PMA output.

It should work for both old style and new style components where the
latter keep a backreference to the catalog item

Also, when a component for example references the same document twice,
it should still be serialized once. This is taken care of in the
pma_spec_experience_page_expPage1.json since the unittest fixture has
two parameters both pointing to the same about-us page

(cherry picked from commit 11ba55282291ec7920c0e2073f6582224d265466)